Implement non-deterministic parser combinators and parser for types.
I have implemented the non-deterministic parser combinators as discussed in the lecture in src/Parser/Combinaters.hs. I have also created a simple Parser for the various types of spl in src/Parser/Parser.hs.

Scanning is done with a a lexer tool (alex) for now. This is just so I could get started with the parser. We will have to change this later. Still, I figure the generated scanner maybe will be handy to test our own scanner against when we make one.