Move precedence information to AST.h
It will probably be needed for the pretty printer as well, so should not be as hardcoded into the parser as it was.

Also add associativity information, which the parser does not use yet. Associativity information is per precedence level, so each precedence level is either right or left associative. I think this is the only way it can make sense, as having two different operators with different associativity on the same precedence level seems to cause problems.