Add parsers for rest of language (declarations)
This completes the majority of the parser. There are still some things that needs to be worked on though. There is a bug in assignments where it does not allow field access in lvals. Also we should add some form of error reporting to the system. Right now finding syntax errors is horrible.