Make the performance of the scanner less glacial.
This is acheived with new eager combinators (like the original we had) that actually gives the performance benefit of eagerness, but sacrifices some error handling capabilities. I think these are error handling capabilities that the scanner does not need. In particular it is fine if the scanner points to the beginning of an identifer or integer literal wich is not welformed instead of pointing to the exact point in the token were to problem is.