Add support for type annotations in Function declarations.
Syntactically this was always the case, but now they work (I think). There are some questions around what the scope of type variables in functions signatures should be.

Right now they are only universally quantified if they do not occur previously (I think). This matches the slides, but I am not sure if it is the most logical thing.

Also, The above should be tested more thuroghly to make sure that is actuallly what is happening.